% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomassF}
\alias{biomassF}
\title{Local biomass summary}
\usage{
biomassF(X, r = NULL, target = NULL, v2 = FALSE, ...)
}
\arguments{
\item{X}{Multitype point pattern of class \code{\link{ppp}} (see package 'spatstat'). The biomass (e.g. size) is to be in an element $mass.}

\item{r}{Vector of sizes for neighbourhoods, e.g. ranges in the \code{geometric} graph neighbourhoods.}

\item{target}{Default NULL. Calculate only for target type. If NULL compute mean over all types.}

\item{v2}{Logical. Return the average biomass instead of just sum (development nomenclature)}

\item{...}{Further parameters for the function \code{\link{segregationFun}}.}
}
\value{
Returns an \code{fv}-object, see \code{spatstat} for more information.
}
\description{
Compute the biomass around points as a function of spatial scale.
}
