% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_tests.R
\name{dixon}
\alias{dixon}
\title{Dixon's 2-type contingency table tests}
\usage{
dixon(X, prepR = 0)
}
\arguments{
\item{X}{bivariate ppp}

\item{prepR}{for large patterns, makes it a bit easier to find nearest neighbours.}
}
\description{
Dixon's 2-type contingency table tests
}
\details{
Computes the tests of segregation using nearest neighbour contingency tables introduced 
by Philip Dixon in his paper "Testing spatial segregation using a nearest-neighbor contingency table",
Ecology, 75, p.1940-1948 (1994). The tests are an improvement on the Pielou's test of segregation.
The test is defined only for two-type spatial pattern.
}
\references{
Philip Dixon: "Testing spatial segregation using a nearest-neighbor contingency table",
Ecology, 75, p.1940-1948 (1994).
}
