% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_tests.R
\name{print.segtest}
\alias{print.segtest}
\title{Print Segtest Objects}
\usage{
\method{print}{segtest}(x, ...)
}
\arguments{
\item{x}{segtest object}

\item{...}{ignored}
}
\description{
Print Segtest Objects
}
