% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{sg.modify.pp}
\alias{sg.modify.pp}
\title{Prepare the input data pattern object}
\usage{
sg.modify.pp(pp)
}
\arguments{
\item{pp}{Point pattern object}
}
\description{
Prepares given point pattern object for computations.
}
